Added dockerfile and documentation to project.
Added documentation for environment variables to server package.
Added a sample environment file and updated the .gitignore to
ignore .env as this file will generally be used to configure the
application locally.
Added a README.md with general usage for the project.
Added a dockerfile for standing up the Mojito API server.