Added candlesticks and a routine to periodically record price data.
Added a package that defines candlesticks which are used to store
and perform math on price data.
Added a goroutine that is optionally invoked when the coinbase
package is loaded. When run it connects to the Coinbase ticker
feed and aggregates and records price data as candlesticks. By
default this occurs every 60 seconds but may be adjusted through
an environment variable. This process may be disabled through an
environment variable as well.