Added an API endpoint for retrieving candlestick data.
Added the exchange field to the candlestick data type.
Added the list candlestick endpoint which allows logged in users
to retrieve price data as candlesticks by exchange and ticker.
Added a package for caching data which contains functions to cache
or retrieve data locally as well as a middleware that will
automatically cache responses and respond with cached data.