Added opens hour and opens day columns to candlestick model.
The candlestick model will now track whether a candlestick opens
a new hour or opens a new day. These columns will be used to
retrieve less dense data as the date range for the get candlesticks
request increases. That is, when we're looking at less than a day
of candlesticks we will see data by the minute, less than two
months of data by the hour, and greater than two months by the
day.