Refactored the price data feed for Coinbase.
Created a market package which will be used populating, tracking,
and retrieving price data.
Added an interface for data feeds that listen to realtime price
data and generate candlestick records.
Rewrote the Coinbase feed so that it implements the feed interface.
Configuration for feeds will now be stored in the database rather
than constants (there is a bit more work to do here but I need to
better define supported exchanges and tickers).
Packages that contain more than one file will now have an init.go
file that consolidates all of the environment variable definitions
and initialization. This will make it easier to understand how a
given package is initialized, previously you would need to look at
each file to see which files defined an init function.